8/10/15 8:45
Fixed the images and put them into the background of whiteboard. Did a
bit of laying out. Please don't remove any comments yet. If you click on
the whiteBoard window the coordinates of where you clicked will be shown
in console as well giving the ability to maybe move objects easier.